+++ cogs/help.py
@@ -25,6 +25,9 @@ class HelpCommand(commands.Cog):
             )
             # MODERATION
             help_embed.add_field(
+                name="!serverinfo", value="Get details about the server.", inline=True
+            )
+            help_embed.add_field(
                 name="!clear", value="Delete a number of messages.", inline=True
             )
             help_embed.add_field(
@@ -109,7 +112,7 @@ class HelpCommand(commands.Cog):
                 inline=True,
             )
             help_embed.set_footer(
-                text=f"Requested by {ctx.author}>.", icon_url=ctx.author.avatar
+                text=f"Requested by {ctx.author}.", icon_url=ctx.author.avatar
             )
 
             await ctx.send(embed=help_embed)
@@ -118,6 +121,7 @@ class HelpCommand(commands.Cog):
             usage_embed = discord.Embed(
                 title="clear",
                 description="Delete some amount of messages in the channel.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage: ", value="!clear <amount>", inline=False)
             usage_embed.add_field(
@@ -126,33 +130,51 @@ class HelpCommand(commands.Cog):
                 inline=True,
             )
             await ctx.send(embed=usage_embed)
+        elif command == "serverinfo":
+            usage_embed = discord.Embed(
+                title="serverinfo",
+                description="Get detailed information about the server.",
+                color=Harbinger.custom_color,
+            )
+            usage_embed.add_field(name="Usage", value="!serverinfo", inline=False)
+            await ctx.send(embed=usage_embed)
         elif command == "join":
             usage_embed = discord.Embed(
-                title="join", description="Join the bot to current voice channel."
+                title="join",
+                description="Join the bot to current voice channel.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage", value="!join", inline=False)
             await ctx.send(embed=usage_embed)
         elif command == "leave":
             usage_embed = discord.Embed(
-                title="leave", description="Remove bot from current voice channel."
+                title="leave",
+                description="Remove bot from current voice channel.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage", value="!leave", inline=False)
             await ctx.send(embed=usage_embed)
         elif command == "pause":
             usage_embed = discord.Embed(
-                title="pause", description="Pause the currently-playing track."
+                title="pause",
+                description="Pause the currently-playing track.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage", value="!pause", inline=False)
             await ctx.send(embed=usage_embed)
         elif command == "play":
             usage_embed = discord.Embed(
-                title="play", description="Resume playback of paused track."
+                title="play",
+                description="Resume playback of paused track.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage", value="!play", inline=False)
             await ctx.send(embed=usage_embed)
         elif command == "stream":
             usage_embed = discord.Embed(
-                title="stream", description="Begin playback of track."
+                title="stream",
+                description="Begin playback of track.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage", value="!stream <URL>", inline=False)
             usage_embed.add_field(
@@ -166,7 +188,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "bug":
             usage_embed = discord.Embed(
-                title="bug", description="Send bot developer(s) a bug report."
+                title="bug",
+                description="Send bot developer(s) a bug report.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!bug <message>", inline=False)
             usage_embed.add_field(
@@ -177,7 +201,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "joined":
             usage_embed = discord.Embed(
-                title="joined", description="Get the datetime a given member joined."
+                title="joined",
+                description="Get the datetime a given member joined.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!joined <member>", inline=False)
             usage_embed.add_field(
@@ -188,7 +214,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "say":
             usage_embed = discord.Embed(
-                title="say", description="Send a message to the channel as the bot."
+                title="say",
+                description="Send a message to the channel as the bot.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(
                 name="Usage:", value="!say <'message'> ", inline=False
@@ -204,7 +232,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "playing":
             usage_embed = discord.Embed(
-                title="playing", description="Send an embed with game info, etc."
+                title="playing",
+                description="Send an embed with game info, etc.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(
                 name="Usage:", value="!playing <game> <field> <value>", inline=False
@@ -232,6 +262,7 @@ class HelpCommand(commands.Cog):
             usage_embed = discord.Embed(
                 title="status",
                 description="Confirm that the bot is online and reachable.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!status", inline=False)
             await ctx.send(embed=usage_embed)
@@ -243,19 +274,25 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "ping":
             usage_embed = discord.Embed(
-                title="ping", description="Get network latency."
+                title="ping",
+                description="Get network latency.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!ping", inline=False)
             await ctx.send(embed=usage_embed)
         elif command == "uptime":
             usage_embed = discord.Embed(
-                title="uptime", description="Get time since bot instance was started."
+                title="uptime",
+                description="Get time since bot instance was started.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!uptime", inline=False)
             await ctx.send(embed=usage_embed)
         elif command == "shutdown":
             usage_embed = discord.Embed(
-                title="shutdown", description="Gracefully shut down the bot."
+                title="shutdown",
+                description="Gracefully shut down the bot.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!shutdown", inline=False)
             await ctx.send(embed=usage_embed)
@@ -263,6 +300,7 @@ class HelpCommand(commands.Cog):
             usage_embed = discord.Embed(
                 title="define",
                 description="Get the Meriam-Webster definition of a word.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!define <word>", inline=False)
             usage_embed.add_field(
@@ -271,7 +309,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "lmgtfy":
             usage_embed = discord.Embed(
-                title="lmgtfy", description="Let Me Google That For You"
+                title="lmgtfy",
+                description="Let Me Google That For You",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!lmgtfy <query>", inline=False)
             usage_embed.add_field(
@@ -284,7 +324,11 @@ class HelpCommand(commands.Cog):
             )
             await ctx.send(embed=usage_embed)
         elif command == "add":
-            usage_embed = discord.Embed(title="add", description="Add two numbers.")
+            usage_embed = discord.Embed(
+                title="add",
+                description="Add two numbers.",
+                color=Harbinger.custom_color,
+            )
             usage_embed.add_field(
                 name="Usage:", value="!add <num1> <num2>", inline=False
             )
@@ -294,7 +338,11 @@ class HelpCommand(commands.Cog):
             usage_embed.add_field(name="<num2: int>", value="Second addend.")
             await ctx.send(embed=usage_embed)
         elif command == "roll":
-            usage_embed = discord.Embed(title="roll", description="Roll some dice.")
+            usage_embed = discord.Embed(
+                title="roll",
+                description="Roll some dice.",
+                color=Harbinger.custom_color,
+            )
             usage_embed.add_field(name="Usage:", value="!roll <NdN>", inline=False)
             usage_embed.add_field(
                 name="<NdN: str>",
@@ -305,7 +353,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "switch":
             usage_embed = discord.Embed(
-                title="switch", description="Turn the Minecraft server on or off."
+                title="switch",
+                description="Turn the Minecraft server on or off.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!switch <on|off>", inline=False)
             usage_embed.add_field(
@@ -318,6 +368,7 @@ class HelpCommand(commands.Cog):
             usage_embed = discord.Embed(
                 title="mccmd",
                 description="Send command(s) to the Minecraft server console.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(name="Usage:", value="!mccmd <command>", inline=False)
             usage_embed.add_field(
@@ -326,7 +377,9 @@ class HelpCommand(commands.Cog):
             await ctx.send(embed=usage_embed)
         elif command == "rps":
             usage_embed = discord.Embed(
-                title="rps", description="Play rock, paper, scissors with the bot."
+                title="rps",
+                description="Play rock, paper, scissors with the bot.",
+                color=Harbinger.custom_color,
             )
             usage_embed.add_field(
                 name="Usage:", value="!rps <rock|paper|scissors>", inline=False