+++ cogs/status.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from email.message import EmailMessage
 import discord
 from discord.ext import commands
-from config.configure import Configure 
+from config.configure import Configure
 
 from harbinger import Harbinger
 
@@ -77,7 +77,7 @@ class Status(commands.Cog):
         file = discord.File(fp="docs/CHANGELOG.md", filename="CHANGELOG.md")
         Harbinger.timestamp(ctx.message.author, cmd, cmd_msg)
         await ctx.send(file=file)
-        
+
     @commands.command()
     async def bug(self, ctx: commands.Context, message) -> None:
         cmd = "!bug"
@@ -87,7 +87,7 @@ class Status(commands.Cog):
         email = EmailMessage()
         email["From"] = "Harbinger"
         email["To"] = email_address
-        email['Subject'] = f'BUG REPORT - {ctx.message.author}'
+        email["Subject"] = f"BUG REPORT - {ctx.message.author}"
         email.set_content(message)
         with smtplib.SMTP(host="smtp.gmail.com", port=587) as smtp:
             smtp.ehlo()
@@ -95,7 +95,9 @@ class Status(commands.Cog):
             smtp.login(email_address, password)
             smtp.send_message(email)
         Harbinger.timestamp(ctx.message.author, cmd, cmd_msg)
-        await ctx.send("Thank you for submitting a bug report.\nIf you'd like to keep abreast of updates/bugfixes, please check out https://github.com/notoriouslogank/harbinger")
+        await ctx.send(
+            "Thank you for submitting a bug report.\nIf you'd like to keep abreast of updates/bugfixes, please check out https://github.com/notoriouslogank/harbinger"
+        )
 
     @commands.command()
     async def bug(self, ctx: commands.Context, message) -> None: