+++ docs/CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - 2023-29-11
+
+### Removed
+
+- Unused function calls in moderation cog
+
+### Added
+
+- serverinfo command to view server details and member list
+
+### Fixed
+
+- Formatting throughout
+
 ## [2.2.1] - 2023-29-11
 
 ### Added